demons/requirements.txt
Django>=4.0.10,<4.1
psycopg2>=2.9.5,<2.10
django-crontab>=0.7.1,<0.8
django-bitfield>=2.2.0,<2.3